Add Author and Comment models with GraphQL schema
Introduces Author and Comment models to the blog app, updates Post to relate to Author, and Comment to relate to Post. Registers new models in the admin, adds migrations, and extends the GraphQL schema to support queries and mutations for authors, posts, and comments. Also adds requirements.txt, updates .gitignore, and provides import/export documentation.